protocols/horizon: dont cast untyped const (#3988)

+++ protocols/horizon/main.go
@@ -96,7 +96,6 @@ func (a Account) GetCreditBalance(code string, issuer string) string {
 // and returns it as a 64-bit integer.
 func (a Account) GetSequenceNumber() (int64, error) {
 	seqNum, err := strconv.ParseInt(a.Sequence, 10, 64)
-
 	if err != nil {
 		return 0, errors.Wrap(err, "Failed to parse account sequence number")
 	}
@@ -112,7 +111,7 @@ func (a *Account) IncrementSequenceNumber() (int64, error) {
 	if err != nil {
 		return 0, err
 	}
-	if seqNum == int64(math.MaxInt64) {
+	if seqNum == math.MaxInt64 {
 		return 0, fmt.Errorf("sequence cannot be increased, it already reached MaxInt64 (%d)", int64(math.MaxInt64))
 	}
 	seqNum++
